# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/import-jobs": {
      "post": {
        "tags": [
          "crate::handlers"
        ],
        "operationId": "enqueue_import_job_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnqueueImportJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Import job enqueued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueImportJobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": [
          "crate::handlers"
        ],
        "operationId": "list_products_handler",
        "responses": {
          "200": {
            "description": "List all products",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProductsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::handlers"
        ],
        "operationId": "create_product_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate handle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/published": {
      "get": {
        "tags": [
          "crate::handlers"
        ],
        "operationId": "list_published_products_handler",
        "responses": {
          "200": {
            "description": "List published products",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProductsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}/publication": {
      "patch": {
        "tags": [
          "crate::handlers"
        ],
        "operationId": "update_product_publication_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductPublicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Product publication updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "crate::handlers"
        ],
        "operationId": "health_handler",
        "responses": {
          "200": {
            "description": "API is healthy",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateProductRequest": {
        "type": "object",
        "required": [
          "title",
          "handle",
          "price_cents",
          "inventory_quantity",
          "published"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "handle": {
            "type": "string"
          },
          "inventory_quantity": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "price_cents": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "published": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "EnqueueImportJobRequest": {
        "type": "object",
        "required": [
          "input_path"
        ],
        "properties": {
          "input_path": {
            "type": "string"
          }
        }
      },
      "EnqueueImportJobResponse": {
        "type": "object",
        "required": [
          "job"
        ],
        "properties": {
          "job": {
            "$ref": "#/components/schemas/ImportJobResponse"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ImportJobResponse": {
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "ListProductsResponse": {
        "type": "object",
        "required": [
          "products"
        ],
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            }
          }
        }
      },
      "ProductResponse": {
        "type": "object",
        "required": [
          "id",
          "title",
          "handle",
          "price_cents",
          "inventory_quantity",
          "published",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "handle": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "inventory_quantity": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "price_cents": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "published": {
            "type": "boolean"
          },
          "published_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "SingleProductResponse": {
        "type": "object",
        "required": [
          "product"
        ],
        "properties": {
          "product": {
            "$ref": "#/components/schemas/ProductResponse"
          }
        }
      },
      "UpdateProductPublicationRequest": {
        "type": "object",
        "required": [
          "published"
        ],
        "properties": {
          "published": {
            "type": "boolean"
          },
          "published_at": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "ahlan-commerce",
      "description": "Ahlan-Commerce API"
    }
  ]
}